backend/requirements.txt
sqlalchemy
fastapi
asyncpg
python-dotenv
uvicorn